# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schedule-slots:
    get:
      operationId: listScheduleSlots
      tags: [schedule]
      summary: List schedule slots in a date range
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Schedule slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleSlot"
    post:
      operationId: createScheduleSlot
      tags: [schedule]
      summary: Create a schedule slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleSlotInput"
      responses:
        "201":
          description: Created schedule slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleSlot"
        "409":
          description: Schedule conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /schedule-slots/generate:
    post:
      operationId: generateSchedule
      tags: [schedule]
      summary: Generate schedule slots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleGenerationInput"
      responses:
        "200":
          description: Generated schedule slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleSlot"
  /schedule-slots/{id}:
    patch:
      operationId: updateScheduleSlot
      tags: [schedule]
      summary: Update a schedule slot
      parameters:
        - $ref: "#/components/parameters/ScheduleSlotId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleSlotUpdate"
      responses:
        "200":
          description: Updated schedule slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleSlot"
        "409":
          description: Locked or conflicting schedule slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteScheduleSlot
      tags: [schedule]
      summary: Delete a schedule slot
      parameters:
        - $ref: "#/components/parameters/ScheduleSlotId"
      responses:
        "204":
          description: Schedule slot deleted
  /act-logs:
    get:
      operationId: listActLogs
      tags: [act-logs]
      summary: List performed diagnostic acts
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: staff_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Performed acts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActLog"
    post:
      operationId: createActLog
      tags: [act-logs]
      summary: Record a performed diagnostic act
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActLogInput"
      responses:
        "201":
          description: Created act log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActLog"
  /act-logs/{id}:
    patch:
      operationId: updateActLog
      tags: [act-logs]
      summary: Update a performed diagnostic act
      parameters:
        - $ref: "#/components/parameters/ActLogId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActLogUpdate"
      responses:
        "200":
          description: Updated act log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActLog"
    delete:
      operationId: deleteActLog
      tags: [act-logs]
      summary: Delete a performed diagnostic act
      parameters:
        - $ref: "#/components/parameters/ActLogId"
      responses:
        "204":
          description: Act log deleted
...
components:
  parameters:
...
    ScheduleSlotId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ActLogId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ShiftType:
      type: string
      enum: [matin, apres_midi, journee_complete, garde_jour, garde_nuit, repos]
    SlotStatus:
      type: string
      enum: [planifie, confirme, annule, remplace]
    ScheduleSlot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        staff_id:
          type: ["string", "null"]
          format: uuid
        shift_date:
          type: string
          format: date
        shift_type:
          $ref: "#/components/schemas/ShiftType"
        status:
          $ref: "#/components/schemas/SlotStatus"
        is_locked:
          type: boolean
        locked_by:
          type: ["string", "null"]
          format: uuid
        locked_at:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, staff_id, shift_date, shift_type, status, is_locked, locked_by, locked_at, notes, created_at, updated_at]
    ScheduleSlotInput:
      type: object
      properties:
        staff_id:
          type: ["string", "null"]
          format: uuid
        shift_date:
          type: string
          format: date
        shift_type:
          $ref: "#/components/schemas/ShiftType"
        status:
          $ref: "#/components/schemas/SlotStatus"
        is_locked:
          type: boolean
        notes:
          type: ["string", "null"]
      required: [shift_date, shift_type]
    ScheduleSlotUpdate:
      type: object
      properties:
        staff_id:
          type: ["string", "null"]
          format: uuid
        shift_date:
          type: string
          format: date
        shift_type:
          $ref: "#/components/schemas/ShiftType"
        status:
          $ref: "#/components/schemas/SlotStatus"
        is_locked:
          type: boolean
        notes:
          type: ["string", "null"]
    ScheduleGenerationInput:
      type: object
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        full_reset:
          type: boolean
      required: [from, to]
    ActLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        staff_id:
          type: string
          format: uuid
        act_id:
          type: string
          format: uuid
        schedule_slot_id:
          type: ["string", "null"]
          format: uuid
        performed_at:
          type: string
          format: date
        quantity:
          type: integer
          minimum: 1
        unit_price:
          type: number
          minimum: 0
        total_amount:
          type: number
          minimum: 0
        entered_by:
          type: ["string", "null"]
          format: uuid
        notes:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, staff_id, act_id, schedule_slot_id, performed_at, quantity, unit_price, total_amount, entered_by, notes, created_at, updated_at]
    ActLogInput:
      type: object
      properties:
        staff_id:
          type: string
          format: uuid
        act_id:
          type: string
          format: uuid
        schedule_slot_id:
          type: ["string", "null"]
          format: uuid
        performed_at:
          type: string
          format: date
        quantity:
          type: integer
          minimum: 1
        entered_by:
          type: ["string", "null"]
          format: uuid
        notes:
          type: ["string", "null"]
      required: [staff_id, act_id, performed_at, quantity]
    ActLogUpdate:
      type: object
      properties:
        staff_id:
          type: string
          format: uuid
        act_id:
          type: string
          format: uuid
        schedule_slot_id:
          type: ["string", "null"]
          format: uuid
        performed_at:
          type: string
          format: date
        quantity:
          type: integer
          minimum: 1
        notes:
          type: ["string", "null"]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error