# lib/api-zod/src/generated/types/actLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActLog {
  id: string;
  staff_id: string;
  act_id: string;
  /** @nullable */
  schedule_slot_id: string | null;
  performed_at: Date;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  unit_price: number;
  /** @minimum 0 */
  total_amount: number;
  /** @nullable */
  entered_by: string | null;
  /** @nullable */
  notes: string | null;
  created_at: Date;
  updated_at: Date;
}