# lib/api-zod/src/generated/types/actLogInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActLogInput {
  staff_id: string;
  act_id: string;
  /** @nullable */
  schedule_slot_id?: string | null;
  performed_at: Date;
  /** @minimum 1 */
  quantity: number;
  /** @nullable */
  entered_by?: string | null;
  /** @nullable */
  notes?: string | null;
}