# lib/api-zod/src/generated/types/scheduleSlotInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShiftType } from './shiftType';
import type { SlotStatus } from './slotStatus';

export interface ScheduleSlotInput {
  /** @nullable */
  staff_id?: string | null;
  shift_date: Date;
  shift_type: ShiftType;
  status?: SlotStatus;
  is_locked?: boolean;
  /** @nullable */
  notes?: string | null;
}