# lib/api-zod/src/generated/types/shiftType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ShiftType = typeof ShiftType[keyof typeof ShiftType];


export const ShiftType = {
  matin: 'matin',
  apres_midi: 'apres_midi',
  journee_complete: 'journee_complete',
  garde_jour: 'garde_jour',
  garde_nuit: 'garde_nuit',
  repos: 'repos',
} as const;