# lib/api-zod/src/generated/types/slotStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type SlotStatus = typeof SlotStatus[keyof typeof SlotStatus];


export const SlotStatus = {
  planifie: 'planifie',
  confirme: 'confirme',
  annule: 'annule',
  remplace: 'remplace',
} as const;